Refactor UI components and implement comprehensive filtering system
- Remove Node Actions panel from graph visualization
- Reposition Node Type legend box for better layout
- Add comprehensive search and filter dropdowns to Table/Card/Kanban views
- Replace simple selects with styled ListView-pattern dropdowns with icons
- Fix Dashboard pie chart legends to use icons instead of circles
- Standardize RightSidebar width across all views (320px)
- Remove radar chart internal titles to prevent duplication
- Clean up unused imports and code
- Remove unused ListView.tsx and TimelineView.tsx files
- Ensure consistent UI patterns and improved user experience
