+++ packages/web/src/components/GraphVisualization.tsx
@@ -448,6 +448,18 @@ export function GraphVisualization() {
           </button>
           <button
             onClick={() => {
+              setSelectedNodeId(undefined);
+              setClickPosition(undefined);
+              setShowCreateModal(true);
+              setNodeMenu(prev => ({ ...prev, visible: false }));
+            }}
+            className="w-full flex items-center px-3 py-2 text-sm text-gray-700 dark:text-gray-200 hover:bg-gray-100 dark:hover:bg-gray-700 transition-colors"
+          >
+            <Plus className="h-4 w-4 mr-2" />
+            Add Node
+          </button>
+          <button
+            onClick={() => {
               setSelectedNodeForEdit(nodeMenu.workItem);
               setShowEditModal(true);
               setNodeMenu(prev => ({ ...prev, visible: false }));
