+++ packages/web/src/components/NodeDistributionRadar.tsx
@@ -121,7 +121,6 @@ export function NodeDistributionRadar({ className = '', showLegend = true }: Nod
       {/* Node Distribution */}
       <div className="bg-gray-800 border border-gray-600 rounded-lg p-6">
         <div className="mb-4">
-          <h3 className="text-lg font-semibold text-white mb-2">Node Distribution</h3>
         </div>
         <div className="relative">
           {/* Zoom Controls */}
