+++ packages/web/src/components/PriorityDistributionRadar.tsx
@@ -123,7 +123,6 @@ export function PriorityDistributionRadar({ className = '', showLegend = true }:
       {/* Priority Distribution */}
       <div className="bg-gray-800 border border-gray-600 rounded-lg p-6">
         <div className="mb-4">
-          <h3 className="text-lg font-semibold text-white mb-2">Priority Distribution</h3>
         </div>
         <div className="relative">
           {/* Zoom Controls */}
